Add time and makefile

+++ GameLogic.h
@@ -15,29 +15,28 @@ const int DIR_DOWN = 1;
 
 class GameLogic
 {
-	int activeLED;
-    int timeToTick;
+    int activeLED;
+    double timeToTick;
     int currentState;
     int direction;
     int maxLeds;
     int startingLED;
     int score;
 
-    ZedBoard * zedboard;
-    Wiimote * wiimote;
+    ZedBoard *zedboard;
+    Wiimote *wiimote;
 
 public:
-	GameLogic(ZedBoard * zed_board, Wiimote * wii_mote, int max_leds, int starting_led) 
-	{
+    GameLogic(ZedBoard *zed_board, Wiimote *wii_mote, int max_leds, int starting_led)
+    {
         zedboard = zed_board;
         wiimote = wii_mote;
-        maxLeds = max_leds - 1; // cause mapping starts from 0
-        startingLED = starting_led - 1; // cause users don't know counting starts from 0 
+        maxLeds = max_leds - 1;         // cause mapping starts from 0
+        startingLED = starting_led - 1; // cause users don't know counting starts from 0
         currentState = STATE_INIT;
     }
 
-
-	bool isAPressed();
+    bool isAPressed();
     void init();
     void start();
     void play();
